refactor: 게시물 생성 및 삭제 기능 추가, 쿼리 파라미터 관리 훅 위치 리팩토링
