--- src/features/post/index.ts
@@ -1,2 +1,0 @@
-export { useUpdatePostMutation } from "../edit-post/api/hooks/use-update-post-mutation"
-export { useDeletePostMutation } from "./api/hooks/use-delete-post-mutation"
