+++ src/widgets/post-dashboard/ui/PostFilter.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { Input, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui"
 import { Search } from "lucide-react"
 import { useGetTags } from "../../../entities/tag/api"
-import { usePostListFilterSearchParams } from "../model/hooks/use-post-list-filter-search-params"
+import { usePostListFilterSearchParams } from "../../../entities/post/model/hooks/use-post-list-filter-search-params"
 
 /**
  * PostFilter
